abstract RangeFinder.java
